+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+## Enhancements
+
+* Adds `Source` and `SourceModuleID` field to `Workspace`, by @jpadrianoGo [#1118](https://github.com/hashicorp/go-tfe/pull/1118)
+
 # v1.81.0
 
 ## Enhancements